% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grimenet-package.R
\docType{package}
\name{grimenet-package}
\alias{grimenet}
\alias{grimenet-package}
\title{grimenet: grimenet}
\description{
Interactive web app for exploring grime music networks.
}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewenhenderson@gmail.com}

}
\keyword{internal}
